1.01

+++ CorvallisBusApp/BusData.h
@@ -49,7 +49,7 @@
 
 @interface BusData : NSObject
 
--(NSArray *)loadArrivalsForLatitude:(float)lat Longitude:(float)lon;
+-(NSArray *)loadArrivalsForLatitude:(CGFloat)lat Longitude:(CGFloat)lon;
 -(NSArray *)loadScheduleForStop:(NSString *)stop;
 
 @property (nonatomic, strong) NSArray *data;
